move plugin config to plugins folder. lazy load nvim-tree
