+++ nvim/init.lua
@@ -1,4 +1,11 @@
-require 'plugins'
+-- bootstrap
+local lazypath = vim.fn.stdpath('data') .. '/lazy/lazy.nvim'
+if not vim.loop.fs_stat(lazypath) then
+  vim.fn.system({ 'git', 'clone', '--filter=blob:none', 'https://github.com/folke/lazy.nvim.git', '--branch=stable', lazypath })
+end
+vim.opt.rtp:prepend(lazypath)
+
+require('lazy').setup('plugins', {})
 require 'settings'
 require 'autocmd'
 require 'commands'
