--- nvim/lua/config/nvim-web-devicons.lua
@@ -1,34 +1,0 @@
-require('nvim-web-devicons').setup {
-  override = {
-    md = {
-      icon = '',
-      color = '#6d8086',
-      name = 'Md'
-    },
-    rb = {
-      icon = '',
-      color = '#e06c75',
-      name = 'Rb'
-    },
-    erb = {
-      icon = '',
-      color = '#e06c75',
-      name = 'Erb',
-    },
-    rake = {
-      icon = '',
-      color = '#e06c75',
-      name = 'Rake'
-    },
-    ['config.ru'] = {
-      icon = '',
-      color = '#e06c75',
-      name = 'ConfigRu'
-    },
-    sqlite3 = {
-      icon = '',
-      color = '#dad8d8',
-      name = 'sqlite',
-    };
-  };
-}
